kafka_pipeline/requirements_kafka.txt
kafka-python==2.2.10
numpy==2.2.6
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
